requirements.txt
flask==3.0.0
gspread==6.0.0
google-auth==2.25.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
google-generativeai>=0.10.0
python-telegram-bot==20.7
rapidfuzz>=3.0.0
